New LoginController in place of the login method in MultiController. Slight introduction to Java annotation based configuration (note that <mvc:annotation-driven /> has been added to libresonic-servlet.xml).
