+++ libresonic-main/src/main/webapp/WEB-INF/libresonic-servlet.xml
@@ -2,7 +2,18 @@
 
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:mvc="http://www.springframework.org/schema/mvc"
+       xsi:schemaLocation="
+        http://www.springframework.org/schema/beans
+        http://www.springframework.org/schema/beans/spring-beans.xsd
+        http://www.springframework.org/schema/context
+        http://www.springframework.org/schema/context/spring-context.xsd
+        http://www.springframework.org/schema/mvc
+        http://www.springframework.org/schema/mvc/spring-mvc.xsd">
+
+    <mvc:annotation-driven />
+    <context:component-scan base-package="org.libresonic.player.controller"/>
 
     <bean id="leftController" class="org.libresonic.player.controller.LeftController">
         <property name="viewName" value="left"/>
@@ -434,7 +445,10 @@
                 <prop key="/top.view">topController</prop>
                 <prop key="/randomPlayQueue.view">randomPlayQueueController</prop>
                 <prop key="/changeCoverArt.view">changeCoverArtController</prop>
+                <!--
+                 login.view has moved to the LoginController
                 <prop key="/login.view">multiController</prop>
+                -->
                 <prop key="/recover.view">multiController</prop>
                 <prop key="/accessDenied.view">multiController</prop>
                 <prop key="/notFound.view">multiController</prop>
